Allocate global data in plasma_data
This is only an issue for bytecode which is assembled into the binary
(what will eventually be cmd.pla, more or less) as loading a module will
automatically allocate its data space in plasma_data. I may just end up
deciding to put all the core VM bytecode in the data space (as the 6502
implementations effectively do) if this proves too difficult - I think
the real test will come when fixing up CALL addresses on loading modules
- but this isn't too hard so far so let's see if we can make it work.

Before this change, global data was effectively being stored at odd
locations in plasma_data (the low 16 bits of the byte address of the
_Dnnnn label in OPC6 space would be used as the address in plasma_data)
but we got away with it because nothing else was living in there.